+++ modules/core/dependency/python-ihm/ihm/dumper.py
@@ -547,14 +547,13 @@ class _EntityPolySegmentDumper(Dumper):
         self._ranges_by_id = []
         # Need to assign ranges for all starting models too
         for sm in system._all_starting_models():
-            rng = sm._get_seq_id_range_all_templates()
+            rng = sm.asym_unit
             util._remove_id(rng, attr='_range_id')
-            sm._all_template_rng = rng
 
         for rng in system._all_entity_ranges():
             util._remove_id(rng, attr='_range_id')
         for rng in itertools.chain(system._all_entity_ranges(),
-                                   (sm._all_template_rng
+                                   (sm.asym_unit
                                     for sm in system._all_starting_models())):
             entity = rng.entity if hasattr(rng, 'entity') else rng
             if entity.is_polymeric():
@@ -1026,7 +1025,7 @@ class _StartingModelDumper(Dumper):
                         entity_id=sm.asym_unit.entity._id,
                         entity_description=sm.asym_unit.entity.description,
                         asym_id=sm.asym_unit._id,
-                        entity_poly_segment_id=sm._all_template_rng._range_id,
+                        entity_poly_segment_id=sm.asym_unit._range_id,
                         starting_model_source=source_map[sm.dataset.data_type],
                         starting_model_auth_asym_id=sm.asym_id,
                         dataset_list_id=sm.dataset._id,
@@ -2479,16 +2478,21 @@ class _FLRPolyProbePositionDumper(Dumper):
     def dump_position(self, system, writer):
         with writer.loop('_flr_poly_probe_position',
                          ['id', 'entity_id', 'entity_description',
+                          'asym_id',
                           'seq_id', 'comp_id', 'atom_id',
                           'mutation_flag', 'modification_flag',
                           'auth_name']) as l:
             for x in self._positions_by_id:
-                comp = x.resatom.entity.sequence[x.resatom.seq_id-1].id
+                comp = x.resatom.entity.sequence[x.resatom.seq_id-1].id if x.resatom.asym is None else x.resatom.asym.entity.sequence[x.resatom.seq_id-1].id
                 atom = None
                 if isinstance(x.resatom, ihm.Atom):
                     atom = x.resatom.id
-                l.write(id=x._id, entity_id=x.resatom.entity._id,
-                        entity_description=x.resatom.entity.description,
+                a_id = None if (x.resatom.asym is None) else x.resatom.asym._id
+                e_id = x.resatom.entity._id if (x.resatom.asym is None) else x.resatom.asym.entity._id
+                e_desc = x.resatom.entity.description if (x.resatom.asym is None) else x.resatom.asym.entity.description
+                l.write(id=x._id, entity_id=e_id,
+                        entity_description=e_desc,
+                        asym_id=a_id,
                         seq_id=x.resatom.seq_id,
                         comp_id=comp, atom_id=atom,
                         mutation_flag=x.mutation_flag,