+++ modules/em/test/medium_test_correlation.py
@@ -56,6 +56,7 @@ class Tests(IMP.test.TestCase):
     def test_simple_correlation(self):
         """ test that the simple fast ccc function works """
         score = self.calc_simple_correlation()
+        print("CCC_score:", score)
         self.assertAlmostEqual(1.000, score, 2)
 
     def test_correlation_with_padding(self):
@@ -159,6 +160,18 @@ class Tests(IMP.test.TestCase):
                                        scores_intervals[i][0],
                                        delta=1e-8)
             self.assertAlmostEqual(result, scores_intervals[i][0], delta=1e-8)
+
+    @IMP.test.expectedFailure
+    def test_correlation_of_added_maps(self):
+        mrw = IMP.em.MRCReaderWriter()
+        ic_mrc = IMP.em.read_map(self.get_input_file_name("inv_crop_map.mrc"), mrw)
+        c_mrc = IMP.em.read_map(self.get_input_file_name("crop_map.mrc"), mrw)
+        mrc = IMP.em.read_map(self.get_input_file_name("1mbn.6.eman.mrc"), mrw)
+
+        ic_mrc.add(c_mrc)
+        ccc =  IMP.em.CoarseCC.cross_correlation_coefficient(mrc, ic_mrc, 0)
+
+        self.assertLess(ccc, 1.00001)
 
 if __name__ == '__main__':
     IMP.test.main()