+++ ChangeLog.md
@@ -1,5 +1,8 @@
 ChangeLog {#changelog}
 =========
+- We no longer provide RPMs for CentOS 6, since it has reached end of life.
+
+# 2.14.0 - 2020-12-02 # {#changelog_2_14_0}
 - All IMP.parallel worker classes are now subclasses of IMP.parallel.Worker;
   the old class names still exist as aliases, but are deprecated.
 - The IMP.kinematics module has had a major overhaul.
@@ -11,6 +14,8 @@ ChangeLog {#changelog}
   - [Integrative threading of protein structure using secondary structure elements](https://salilab.org/SSEThread/)
 - Bugfix: fix near-infinite loop in FoXS JavaScript output when no models
   are fit.
+- Bugfix: IMP.saxs.Restraint should now work correctly when some or all
+  particles are members of rigid bodies.
 
 # 2.13.0 - 2020-05-14 # {#changelog_2_13_0}
 - Ubuntu .deb packages now use Python 3 by default (install the imp-python2